refactor: move commands into it's own module
The structure for commands needs to be scalable as it's the most common
addition to the bot. Move commands into it's own directory instead of
maintaining one file.
